requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pyserial>=3.5
flask>=3.0.0
flask-socketio>=5.3.0
sounddevice>=0.4.6
